Renamed LogEntryFields.
The names are now consistent with the rest of the program, and there is no more confusion whether fid is the ID of the logfile or the reader, or whether file is the logfile or the source file.

The keys in the JSON dict pushed to Redis remain unaltered, of course.